restructuring of component folder, file renames, removal of unnecessary files
